removed redundant import, moved files around
