+++ sf_examples/envpool_examples/mujoco/envpool_mujoco_utils.py
@@ -5,36 +5,14 @@ try:
 except ImportError as e:
     print(e)
     print("Trying to import envpool when it is not install. install with 'pip install envpool'")
+    raise e
 
-import gym
-
-from sample_factory.utils.utils import is_module_available
+# from sample_factory.utils.utils import is_module_available
 from sf_examples.envpool_examples.envpool_wrappers import EnvPoolTo5Tuple
+from sf_examples.mujoco_examples.mujoco.mujoco_utils import MUJOCO_ENVS
 
-
-def mujoco_available():
-    return is_module_available("mujoco")
-
-
-class MujocoSpec:
-    def __init__(self, name, env_id):
-        self.name = name
-        self.env_id = env_id
-
-
-MUJOCO_ENVS = [
-    MujocoSpec("mujoco_hopper", "Hopper-v4"),
-    MujocoSpec("mujoco_halfcheetah", "HalfCheetah-v4"),
-    MujocoSpec("mujoco_humanoid", "Humanoid-v4"),
-    MujocoSpec("mujoco_ant", "Ant-v4"),
-    MujocoSpec("mujoco_standup", "HumanoidStandup-v4"),
-    MujocoSpec("mujoco_doublependulum", "InvertedDoublePendulum-v4"),
-    MujocoSpec("mujoco_pendulum", "InvertedPendulum-v4"),
-    MujocoSpec("mujoco_reacher", "Reacher-v4"),
-    MujocoSpec("mujoco_walker", "Walker2d-v4"),
-    MujocoSpec("mujoco_pusher", "Pusher-v4"),
-    MujocoSpec("mujoco_swimmer", "Swimmer-v4"),
-]
+# def mujoco_available():
+#     return is_module_available("mujoco")
 
 
 def mujoco_env_by_name(name):
