+++ sf_examples/envpool_examples/mujoco/train_envpool_mujoco.py
@@ -3,11 +3,9 @@ import sys
 from sample_factory.cfg.arguments import parse_full_cfg, parse_sf_args
 from sample_factory.envs.env_utils import register_env
 from sample_factory.train import run_rl
-from sf_examples.envpool_examples.mujoco_examples.mujoco.envpool_mujoco_params import (
-    add_mujoco_env_args,
-    mujoco_override_defaults,
-)
-from sf_examples.envpool_examples.mujoco_examples.mujoco.envpool_mujoco_utils import MUJOCO_ENVS, make_mujoco_env
+from sf_examples.envpool_examples.mujoco.envpool_mujoco_params import add_mujoco_env_args, mujoco_override_defaults
+from sf_examples.envpool_examples.mujoco.envpool_mujoco_utils import make_mujoco_env
+from sf_examples.mujoco_examples.mujoco.mujoco_utils import MUJOCO_ENVS
 
 
 def register_mujoco_components():
